services/reader/requirements.txt
aiohttp==3.12.15
aiohttp-sse==2.2.0
python-dotenv==1.1.1
